@@ -25,31 +25,29 @@ apply plugin: 'com.android.library'
 //}
 
 android {
-  compileSdkVersion 23
-  buildToolsVersion "23.0.3"
+    compileSdkVersion 28
 
-  defaultConfig {
-    minSdkVersion 16
-    targetSdkVersion 23
-    versionCode 2
-    versionName "1.4"
-  }
-  buildTypes {
-    release {
-      minifyEnabled false
+    defaultConfig {
+        minSdkVersion 16
+        targetSdkVersion 28
     }
-  }
+    buildTypes {
+        release {
+            minifyEnabled false
+        }
+    }
+    namespace 'com.thoughtbot.expandablerecyclerview'
 }
 
 dependencies {
-  //android
-  compile 'com.android.support:appcompat-v7:23.4.0'
-  compile 'com.android.support:recyclerview-v7:23.4.0'
+    //android
+    implementation 'androidx.appcompat:appcompat:1.5.1'
+    implementation 'androidx.recyclerview:recyclerview:1.2.1'
 
-  //unit tests
-  testCompile 'junit:junit:4.12'
-  testCompile 'org.robolectric:robolectric:3.2.1'
-  testCompile 'org.mockito:mockito-core:1.10.5'
+    //unit tests
+    testImplementation 'junit:junit:4.12'
+    testImplementation 'org.robolectric:robolectric:3.2.1'
+    testImplementation 'org.mockito:mockito-core:1.10.5'
 }
 //apply from: 'https://raw.githubusercontent.com/nuuneoi/JCenter/master/installv1.gradle'
 //apply from: 'https://raw.githubusercontent.com/nuuneoi/JCenter/master/bintrayv1.gradle'
