@@ -11,10 +11,12 @@ import java.util.List;
 public class ExpandableGroup<T extends Parcelable> implements Parcelable {
   private String title;
   private List<T> items;
+  private int viewType;
 
-  public ExpandableGroup(String title, List<T> items) {
+  public ExpandableGroup(String title, List<T> items, int viewType) {
     this.title = title;
     this.items = items;
+    this.viewType = viewType;
   }
 
   public String getTitle() {
@@ -64,10 +66,8 @@ public class ExpandableGroup<T extends Parcelable> implements Parcelable {
     } else {
       dest.writeByte((byte) (0x01));
       dest.writeInt(items.size());
-      if (items.size() > 0) {
-        final Class<?> objectsType = items.get(0).getClass();
-        dest.writeSerializable(objectsType);
-      }
+      final Class<?> objectsType = items.get(0).getClass();
+      dest.writeSerializable(objectsType);
       dest.writeList(items);
     }
   }
@@ -85,4 +85,8 @@ public class ExpandableGroup<T extends Parcelable> implements Parcelable {
           return new ExpandableGroup[size];
         }
       };
+
+  public int getViewType() {
+    return viewType;
+  }
 }
